--- src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Archer.java
@@ -1,7 +1,0 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
-
-public class Archer extends Hero {
-    public Archer(String name) {
-        super(name);
-    }
-}
