--- src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Mag.java
@@ -1,7 +1,0 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
-
-public class Mag extends Hero {
-    public Mag(String name) {
-        super(name);
-    }
-}
