--- src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Warrior.java
@@ -1,7 +1,0 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
-
-public class Warrior extends Hero {
-    public Warrior(String name) {
-        super(name);
-    }
-}
