+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/heroes/Hero.java
@@ -1,4 +1,6 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
+package com.rakovets.course.java.core.practice.oop_principles.battle_ground.heroes;
+
+import com.rakovets.course.java.core.practice.oop_principles.battle_ground.Enemy;
 
 public class Hero {
     private String name;
