+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/cat/cats/Sphynx.java
@@ -1,6 +1,9 @@
-package com.rakovets.course.java.core.practice.oop_principles.cat;
+package com.rakovets.course.java.core.practice.oop_principles.cat.cats;
 
-public class Sphynx extends Cat{
+import com.rakovets.course.java.core.practice.oop_principles.cat.Person;
+import com.rakovets.course.java.core.practice.oop_principles.cat.cats.Cat;
+
+public class Sphynx extends Cat {
 
     public Sphynx(String name) {
         super(name);
